handle exception and config response

+++ src/main/java/com/aps/projectmanage/service/impl/UserServiceImpl.java
@@ -1,41 +1,58 @@
 package com.aps.projectmanage.service.impl;
 
+import com.aps.projectmanage.domain.dto.UserDTO;
 import com.aps.projectmanage.domain.entity.User;
 import com.aps.projectmanage.domain.repository.UserRepository;
+import com.aps.projectmanage.exception.GlobalExceptionHandler;
+import com.aps.projectmanage.mapper.UserMapper;
 import com.aps.projectmanage.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
+    private final UserMapper userMapper;
 
     @Override
-    public User createUser(User user) {
-        return userRepository.save(user);
+    public UserDTO create(UserDTO userDTO) {
+        User user = userMapper.toEntity(userDTO);
+        user = userRepository.save(user);
+        return userMapper.toDTO(user);
     }
 
     @Override
-    public Optional<User> getUserById(Long id) {
-        return userRepository.findById(id);
+    public Optional<UserDTO> getById(Long id) {
+        return userRepository.findById(id)
+                .map(userMapper::toDTO);
     }
 
     @Override
-    public List<User> getAllUsers() {
-        return userRepository.findAll();
+    public List<UserDTO> getAll() {
+        return userRepository.findAll()
+                .stream()
+                .map(userMapper::toDTO)
+                .collect(Collectors.toList());
     }
 
     @Override
-    public User updateUser(User user) {
-        return userRepository.save(user);
+    public UserDTO update(UserDTO userDTO, Long id) {
+        User existingUser = userRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("User not found with id: " + id));
+
+        return userMapper.toDTO(userRepository.save(existingUser));
     }
 
     @Override
-    public void deleteUser(Long id) {
+    public void deleteById(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new RuntimeException("User not found with id: " + id);
+        }
         userRepository.deleteById(id);
     }
 }
